Removing the deprecated CatalogController#admin?; Removing the deprecated Blacklight::ConfigurationHelperBehavior#sort_fields in favor of #active_sort_fields; Correcting the Capybara #have_link assertion within the Scanned Resources Feature test suite (this was logging a warning; Removing the SolrFacadeService and directly instantiating SolrFacade Objects
