+++ app/views/catalog/_members_collection.html.erb
@@ -5,7 +5,7 @@
       <fieldset class="col-xs-12">
         <legend class="sr-only"><%= t('sort_label') %></legend>
         <%= label_tag(:sort, "<span>Sort By:</span>".html_safe) %>
-        <%= select_tag(:sort, options_for_select(sort_fields, h(params[:sort]))) %>
+        <%= select_tag(:sort, options_for_select(active_sort_fields, h(params[:sort]))) %>
         <%= label_tag(:per_page) do %>
           Show <%= select_tag(:per_page, options_for_select(['10', '20', '50', '100'], h(params[:per_page])), title: "Number of results to display per page") %>
           per page
